Adding QueryExecutor. Adding libraries gflags and glog; adding logging for descriptors.

+++ be/src/runtime/mem-pool.cc
@@ -23,7 +23,7 @@ MemPool::MemPool(int chunk_size)
 }
 
 MemPool::~MemPool() {
-  for (int i = 0; i < mem_chunks_.size(); ++i) {
+  for (size_t i = 0; i < mem_chunks_.size(); ++i) {
     delete [] mem_chunks_[i];
   }
 }
